abstract scrollable table to reusable component
