+++ client/src/components/ScrollableTableWithMenu/index.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import ScrollableTableWithMenu from './index';
+
+describe( '<ScrollableTableWithMenu />', () => {
+  it( 'renders without crashing', () => {
+    shallow( <ScrollableTableWithMenu /> );
+  } );
+} );
