+++ client/src/containers/Admin/Dashboard/MyProjects/constants.js
@@ -18,48 +18,77 @@ export const menuItems = [
 
 export const tempData = [
   {
+    id: 1,
     name: 'John',
     status: 'Approved',
     notes: 'None',
     analytics: '1'
   },
   {
+    id: 2,
     name: 'Jamie',
     status: 'Approved',
     notes: 'Requires call',
     analytics: '2'
   },
   {
+    id: 3,
     name: 'John',
     status: 'Approved',
     notes: 'None',
     analytics: '3'
   },
   {
+    id: 4,
     name: 'Jamie',
     status: 'Approved',
     notes: 'Requires call',
     analytics: '4'
   },
   {
+    id: 5,
     name: 'John',
     status: 'Approved',
     notes: 'None',
     analytics: '5'
   },
   {
+    id: 6,
     name: 'Jamie',
     status: 'Approved',
     notes: 'Requires call',
     analytics: '6'
   },
   {
+    id: 7,
     name: 'John',
     status: 'Approved',
     notes: 'None',
     analytics: '7'
   },
   {
+    id: 8,
+    name: 'Jamie',
+    status: 'Approved',
+    notes: 'Requires call',
+    analytics: '8'
+  },
+  {
+    id: 9,
+    name: 'Jamie',
+    status: 'Approved',
+    notes: 'Requires call',
+    analytics: '6'
+  },
+  {
+    id: 10,
+    name: 'John',
+    status: 'Approved',
+    notes: 'None',
+    analytics: '7'
+  },
+  {
+    id: 11,
     name: 'Jamie',
     status: 'Approved',
     notes: 'Requires call',
