+++ client/src/containers/Admin/Dashboard/MyProjects/index.js
@@ -9,165 +9,21 @@ import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 import * as actions from './actions';
 import makeSelectMyProjects from './selectors';
-import { Table, Grid, Checkbox, Icon, Label } from 'semantic-ui-react';
 import './MyProjects.css';
-
-import lowerFirst from 'lodash/lowerFirst';
-import upperFirst from 'lodash/upperFirst';
-import sortBy from 'lodash/sortBy';
-
 import { tempData, menuItems } from './constants';
+import ScrollableTableWithMenu from 'components/ScrollableTableWithMenu';
 
 /* eslint-disable react/prefer-stateless-function */
 class MyProjects extends React.Component {
-  state = {
-    displayTableMenu: false,
-    tableHeaders: ['name', 'status', 'notes'],
-    selectAllProjects: false,
-    selectedProjects: new Map(),
-    column: null,
-    data: tempData,
-    direction: null
-  };
-
-  toggleTableMenu = () => {
-    this.setState( prevState => ({ displayTableMenu: !prevState.displayTableMenu }) );
-  }
-
-  tableMenuOnChange = ( e ) => {    
-    e.persist();
-    const menuItem = lowerFirst(e.target.textContent);
-    this.setState(prevState => {
-      if ( prevState.tableHeaders.includes(menuItem) ) {
-        return { tableHeaders: prevState.tableHeaders.filter( header => header !== menuItem ) };
-      } else {
-        return { tableHeaders: [...prevState.tableHeaders, menuItem] };
-      }
-    });
-  }
-
-  toggleAllCheckboxSelection = () => {    
-    const allProjects = Array
-      .from( document.querySelectorAll('[data-label]') )
-      .map( project => project.dataset.label );
-
-    const newSelectAllProjectsState = !this.state.selectAllProjects;
-    let newSelectedProjects = new Map();        
-    
-    allProjects.map( project => {
-      if ( !newSelectAllProjectsState ) {
-        newSelectedProjects.set(project, false);
-      } else {
-        newSelectedProjects.set(project, true);
-      }      
-    } );
-    
-    this.setState({
-      selectAllProjects: newSelectAllProjectsState,
-      selectedProjects: newSelectedProjects
-    });
-  }
-
-  toggleProjectSelection = (e, data) => {
-    const isChecked = data.checked;
-    this.setState( prevState => ( {
-      selectedProjects: prevState.selectedProjects.set(data['data-label'], isChecked)
-    } ) );
-  }
-
-  handleSort = clickedColumn => () => {
-    console.log(clickedColumn);
-    const { column, data, direction } = this.state
-
-    if (column !== clickedColumn) {
-      this.setState({
-        column: clickedColumn,
-        data: sortBy(data, [clickedColumn]),
-        direction: 'ascending',
-      });
-      return;
-    }
-
-    this.setState({
-      data: data.reverse(),
-      direction: direction === 'ascending' ? 'descending' : 'ascending',
-    });
-  }
-
   render() {
-    const {
-      displayTableMenu,
-      tableHeaders,
-      selectAllProjects,
-      selectedProjects,
-      column,
-      direction
-    } = this.state;
+    const tableHeaders = ['name', 'status', 'notes'];
 
     return (
-      <Grid>
-        <Grid.Row>
-          <Grid.Column floated='left' width={ 8 }>
-            <div>Show: 25 &#9660; | 1-25 of 137 for 'search term'</div>
-          </Grid.Column>
-          <Grid.Column floated='right' width={ 8 }>
-            <div className='myProjects_menu'>
-              <span onClick={ this.toggleTableMenu }>See More &#9660;</span>
-              <div className={ displayTableMenu ? 'myProjects_menu_list display' : 'myProjects_menu_list' }>
-                { menuItems.map( item => (
-                  <Checkbox key={ item } onChange={ this.tableMenuOnChange } label={ item } />
-                ) ) }
-              </div>
-            </div>
-          </Grid.Column>
-        </Grid.Row>        
-        <Grid.Row>          
-          <Grid.Column>
-            <div className="myProjects_table">
-              <Table sortable>
-                <Table.Header>
-                  <Table.Row>
-                    { tableHeaders.map( (header,i) => (
-                      <Table.HeaderCell 
-                        key={i}
-                        className="myProjects_table_headerCell"
-                        sorted={ column === header ? direction : null }
-                        onClick={ this.handleSort( header ) }
-                      >
-                        { i === 0 
-                          ? <Checkbox label={ upperFirst(header) } onClick={ this.toggleAllCheckboxSelection }/>
-                          : upperFirst(header)
-                        }
-                      </Table.HeaderCell>
-                    ) ) }
-                  </Table.Row>
-                </Table.Header>
-                <Table.Body>
-                  { tempData.map( (d,i) => (
-                    <Table.Row key={`${d.name}_${i}`}>
-                      { tableHeaders.map( header => {
-                        return (
-                          <Table.Cell key={header}>
-                            { header === 'name' 
-                              ? <Checkbox
-                                  data-label={ `name_${d.name}_${i}` }
-                                  label={ d[header] }
-                                  checked={ selectedProjects.get(`name_${d.name}_${i}`) }
-                                  onChange={ this.toggleProjectSelection }
-                                /> 
-                              : d[header]
-                            }
-                          </Table.Cell>
-                        )
-                      } ) }
-                    </Table.Row>
-                  ) ) }
-                </Table.Body>
-              </Table>
-            </div>
-          </Grid.Column>
-        </Grid.Row>
-      </Grid>
+      <ScrollableTableWithMenu
+        tableData={ tempData }
+        columnMenu={ menuItems }
+        persistentTableHeaders={ tableHeaders }
+      />
     );
   }
 }
